// src/infrastructure/external-services/ai-generate-image/comfyui/upscale/types/NodeUpscaleInfo.ts
export enum NodeUpscaleInfo {
  UPSCALE_IMAGE_WITH_MODEL_ID = '%%image_upscale_with_model_id%%',
  UPSCALE_MODEL_LOADER_ID = '%%upscale_model_loader_id%%',
  LOAD_IMAGE_ID = '%%load_image_id%%',
  LOAD_IMAGE_FILENAME = '%%load_image_filename%%',
  SAVE_IMAGE_ID = '%%save_image_id%%',
  SAVE_IMAGE_FILENAME = '%%save_image_filename%%',
  UPSCALE_MODEL_LOADER_MODEL_NAME = '%%upscale_model_loader_model_name%%',
  UPSCALE_MODEL_LOADER_TITLE = 'UpscaleModelLoader_Upscale',
  UPSCALE_IMAGE_WITH_MODEL_TITLE = 'ImageUpscaleWithModel_Upscale',
}